refactor: allow requests and responses to be extended independently
